chore: move compiler logic out of index.js
